Fix lifetime of `ExclusiveCanvas` canvas references
To get mutable access to a `Canvas` reference you must first call
`gui_direct_draw_acquire` then release it with `gui_direct_draw_release`
when done.

In our Rust wrapper, this is enforced by returning an `ExclusiveCanvas`
handle which represents the claim and can be use for borrowing a `Canvas`.

As such, the following should not compile (temporary value dropped while borrowed):

```rust
let canvas: &mut Canvas = gui.direct_draw_acquire().canvas_mut();
````

However, before this fix it did, due to the borrow happening on `Gui`
and not `ExclusiveCanvas`. This would make it impossible to actually
write to the Canvas, which was part of the issue in #235.

`ExclusiveCanvas` uses `PhantomData` to hold a "reference" to `Gui`,
so by holding a reference to `ExclusiveCanvas` we transitively also
hold a reference to `Gui`, thus don't need to explicitly include `Gui`'s
lifetime in returned references.